requirements.txt
pyswip>=0.2.10
streamlit>=1.31.0
openai>=1.12.0
pyyaml>=6.0.1
networkx>=3.2.1
numpy>=1.26.0
pytest>=8.0.0
python-dotenv>=1.0.0
zmq>=0.0.0
pydantic>=2.6.0
fastapi>=0.109.0
uvicorn>=0.27.0 